rename components to match naming schema
