+++ src/components/molecules/FabIcon.unit.js
@@ -1,0 +1,5 @@
+import FabIcon from "./FabIcon";
+
+describe("@components/FabIcon", () => {
+	it(...isValidComponent(FabIcon));
+});
