--- src/components/molecules/IconFab.unit.js
@@ -1,5 +1,0 @@
-import IconFab from "./IconFab";
-
-describe("@components/IconFab", () => {
-	it(...isValidComponent(IconFab));
-});
